@@ -1,15 +1,34 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.0.10'
-gem 'bson_ext'
-gem 'mongo_mapper'
-gem 'jquery-rails'
-gem 'compass'
-gem 'haml'
-gem 'haml-rails'
+gem 'rails', '3.1.0'
 
-# Bundle edge Rails instead:
-# gem 'rails', :git => 'git://github.com/rails/rails.git'
+
+gem 'rails', '3.1.0'
+group :assets do
+  gem 'sass-rails', "  ~> 3.1.0"
+  gem 'coffee-rails', "~> 3.1.0"
+  gem 'uglifier'
+  gem 'compass', '~> 0.12.alpha'
+end
+gem 'jquery-rails'
+gem "rspec-rails", ">= 2.6.1", :group => [:development, :test]
+gem "database_cleaner", ">= 0.6.7", :group => :test
+gem "mongoid-rspec", ">= 1.4.4", :group => :test
+gem "factory_girl_rails", ">= 1.1.0", :group => :test
+# gem "cucumber-rails", ">= 1.0.2", :group => :test
+gem "capybara", ">= 1.0.1", :group => :test
+gem "launchy", ">= 2.0.5", :group => :test
+gem "bson_ext", ">= 1.3.1"
+gem "mongoid", ">= 2.2.0"
+gem "omniauth", ">= 0.2.6"
+gem 'devise'
+gem 'haml'
+gem 'haml-rails', :group => :development
+gem 'delayed_job'
+gem 'delayed_job_mongoid'
+gem 'linkedin'
+gem 'yettings'
+
 
 
 # Use unicorn as the web server
@@ -18,21 +37,6 @@ gem 'haml-rails'
 # Deploy with Capistrano
 # gem 'capistrano'
 
-# To use debugger (ruby-debug for Ruby 1.8.7+, ruby-debug19 for Ruby 1.9.2+)
-# gem 'ruby-debug'
+# To use debugger
 # gem 'ruby-debug19', :require => 'ruby-debug'
 
-# Bundle the extra gems:
-# gem 'bj'
-# gem 'nokogiri'
-# gem 'sqlite3-ruby', :require => 'sqlite3'
-# gem 'aws-s3', :require => 'aws/s3'
-
-# Bundle gems for the local environment. Make sure to
-# put test-only gems in this group so their generators
-# and rake tasks are available in development mode:
-group :development, :test do
-  gem "rspec-rails"
-  gem "database_cleaner"
-  # gem 'webrat'
-end
