@@ -25,7 +25,14 @@ RSpec.configure do |config|
   # instead of true.
   # config.use_transactional_fixtures = true
   
+  # Clean up the database
+  require 'database_cleaner'
+  config.before(:suite) do
+    DatabaseCleaner.strategy = :truncation
+    DatabaseCleaner.orm = "mongoid"
+  end
+
   config.before(:each) do
-    DatabaseCleaner.clean_with(:truncation)
+    DatabaseCleaner.clean
   end
 end
